[ ADD ] dividends_fund table & fix types of relation ids

+++ supabase/migrations/20230728221131_create_company_events_table.sql
@@ -6,7 +6,7 @@ create table
     title text,
     description text,
     timestamp bigint,
-    stock_id integer references stocks,
+    stock_id bigint references stocks,
     created_at timestamp with time zone default timezone('utc'::text, now()) not null,
     updated_at timestamp with time zone default timezone('utc'::text, now()) not null
 );
